Tickets:SLSUP-2, SLSUP-4, SLSUP-5 1. Decimal usage id to usage name mapping 2. Configured project easier to be built and pulled 3. 'salesup' project now has two applications inside: cloud & local
