+++ .gitignore
@@ -39,6 +39,7 @@ cmake-build-debug/
 
 # IntelliJ
 /out/
+**/out/
 
 # mpeltonen/sbt-idea plugin
 .idea_modules/
@@ -63,10 +64,10 @@ fabric.properties
 # See https://github.com/joeblau/gitignore.io/issues/186 and https://github.com/joeblau/gitignore.io/issues/360
 
 .idea/
-*.iml
-
 
 # End of https://www.gitignore.io/api/intellij+all
+
+
 
 #####################
 #### Gradle part ####
