+++ apps/cloud/src/main/java/com/xzxd/salesup/CloudApplication.java
@@ -1,0 +1,11 @@
+package com.xzxd.salesup;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CloudApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CloudApplication.class, args);
+    }
+}
