+++ apps/cloud/src/main/java/com/xzxd/salesup/repositories/SalesRecordRepository.java
@@ -1,0 +1,7 @@
+package com.xzxd.salesup.repositories;
+
+import com.xzxd.salesup.domains.SalesRecord;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SalesRecordRepository extends JpaRepository<SalesRecord, Long> {
+}
