+++ apps/cloud/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+server.port=0
+
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.url=jdbc:mysql://localhost:3306/salesup?useSSL=false
+spring.datasource.username=root
+spring.datasource.password=
