+++ build.gradle
@@ -1,46 +1,34 @@
 buildscript {
     repositories {
-        mavenCentral()
+        mavenLocal()      // ~/.m2/repoitory/
+        mavenCentral()    // https://repo1.maven.org/maven2/
     }
     dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.0.0.RELEASE")
     }
 }
 
-apply plugin: 'java'
-apply plugin: 'eclipse'
-apply plugin: 'idea'
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
+allprojects {
+    apply plugin: 'idea'
 
-repositories {
-    mavenCentral()
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+subprojects {
+    apply plugin: 'maven' // gradle install
 
-dependencies {
-    compile("org.springframework.boot:spring-boot-starter-web")
-    testCompile('org.springframework.boot:spring-boot-starter-test')
+    afterEvaluate {
+        if (project.plugins.hasPlugin('java')) {
+            sourceCompatibility = JavaVersion.VERSION_1_8
+        }
+    }
 
-    /*
-    Data
-     */
-    // JPA Data (We are going to use Repositories, Entities, Hibernate, etc...)
-    compile 'org.springframework.boot:spring-boot-starter-data-jpa'
-    // Use MySQL Connector-J
-    compile 'mysql:mysql-connector-java'
+    repositories {
+        mavenLocal()
+        mavenCentral()
+    }
 
-    /*
-    Security
-    */
-    compile("org.springframework.boot:spring-boot-starter-security")
-    compile("org.springframework.security.oauth:spring-security-oauth2:2.2.1.RELEASE")
 
-    /*
-    Keyboard hook
-     */
-    compile("com.1stleg:jnativehook:2.1.0")
-    compile("org.hid4java:hid4java:0.5.0")
+    // gradle publish
+    // ~/.m2/repoitory/com/company/project/<*>/0.0.1/<prod>-0.0.1.jar
+    project.group = 'com.xzxd.salesup'
+    project.version = '0.0.1'
 }
