+++ gradle/wrapper/gradle-wrapper.properties
@@ -3,4 +3,4 @@ distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://com.XZXD.salesup.services.gradle.org/distributions/gradle-2.9-bin.zip
+distributionUrl=http\://services.gradle.org/distributions/gradle-4.6-all.zip
