--- src/main/java/com/XZXD/salesup/controllers/QueryController.java
@@ -1,23 +1,0 @@
-package com.XZXD.salesup.controllers;
-
-import com.XZXD.salesup.domains.SalesRecord;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import com.XZXD.salesup.services.QueryService;
-
-import java.util.List;
-
-
-@RestController
-public class QueryController {
-
-    @Autowired
-    QueryService queryService;
-
-    @RequestMapping("/query/all/sales_record")
-    public List<SalesRecord> queryAllSalesRecord() {
-        return queryService.queryAllSalesRecord();
-    }
-
-}
